Make performance counters always 64 bits wide (#178)
This prevents the counters from easily overflowing for a 32-bit pk.

+++ pk/boot.h
@@ -20,9 +20,9 @@ typedef struct {
   size_t brk_max;
   size_t mmap_max;
   size_t stack_top;
-  size_t time0;
-  size_t cycle0;
-  size_t instret0;
+  uint64_t time0;
+  uint64_t cycle0;
+  uint64_t instret0;
 } elf_info;
 
 extern elf_info current;
